Manage allowed upload users via artisan console (#12)
This change adds two custom artisan commands: uploadUsers:show and uploadUsers:set {file}. The first one shows the list of users who are allowed to upload mismatch files to the store. The second one takes the name of an allowlist file as an argument and replaces the current list of users with the one from the specified file. The allow list file must be placed in the storage/app/allowlist directory. Users who are sending API requests to the /api/upload/ endpoint will consequently be checked against this list.

* Check file uploads against user allow list

* Add a description to the 'allowlist' input argument

* Document allow list management in Admin Guide

* Add uploader seeder

* Create authorization gate for uploads

Co-authored-by: Itamar Givon <itamar.givon@wikimedia.de>
Co-authored-by: Guergana Tzatchkova <guergana.tzatchkova@wikimedia.de>

Bug: T286024