Use rsync to deploy, don't delete and re-copy
In order for this to work reliably, we make sure that
the deployed files are group-writable, then change the
new source files to have tools.mismatch-finder as group.
After that, we rsync the new files in place.
Finally, we run take to make sure all (new) files/folders
are owned by tools.mismatch-finder. Due to take's symlink
handling, we need to run it on for all files/folders, and
can't let it recurse on its own.

https://phabricator.wikimedia.org/T286056