Update mismatch migration (#35)
This PR corrects the mismatch migration to fit the most minimal product requirements, keeping in mind length limitations. As it turns out, since we do not intend to do anything with the wikidata value besides showing it, there is no need to save it as a BLOB. In addition, since we limit the length of external values to 1500 then we do not need to store it as TEXT. In case any of these have to be updated in the future, we can create an additional migration to handle that.

Bug: T285299