[Imports 3/7] Create Wikidata API Client (#38)
This change introduces a Wikibase API Client Service, so that we can call Wikidata API endpoints and cache our responses. The service already includes a method to call `wbparsevalue` which throws a `WikibaseValueParserException` on errors. Additionally, we wire up this service to the service container in `WikidataAPIProvider`, as it has external dependencies, and thus cannot be injected with Laravel's Zero-configuration resolution mechanism.

This change also installs some caching middleware, in order to enable us to cache responses for our API calls to try to keep our call rate to a minimum.

Bug: T285299