Re-create mismatches table instead of ALTERing it
This change modifies a database migration, which adds an 'item_id'
column to the mismatches table. Since SQLite does not allow adding
non-nullable columns without a default value, the cleanest way is to
drop the entire table and re-create it with the new column. No
mismatches will be harmed, if this is executed immediately after
creating the table using 2021_07_19_123858_create_mismatches_table.php

Bug: T285300