Update the Mismatch concept (#75)
This modifies the Mismatch model to include a reviewer and renames the
'status' field to 'review_status'.

For these changes to take effect, several components had to be touched:
 * the Mismatch model,
 * migrations for the 'mismatches' database table,
 * the factory, controller and resource,
 * the API route test.

Since Laravel currently does not support renaming an 'enum' column, the
db migration drops and re-creates it at the same position with 'pending'
as its default value. This will do no harm, because no review decisions
have been recorded so far. SQLite does not even allow dropping and
adding columns, therefore in the 'testing' environment, the entire
table will be dropped and re-created with the new column name.

Bug: T289132