Add querying section to home page (#82)
This change adds the querying section to the home page of the Mismatch Finder application. This section contains a form to send item IDs to the server, which will be used to filter mismatches to retrieve. The submit button is currently disabled, as we cannot yet display the results of the form.

* Add messages
* Add typographic mixins
* Add initial page markup and state logic
* Apply typography and layout styles
* Apply page specific styles
* Add browser test to cover BDD specs

Bug: T289057