Implement app loading state (#94)
This change introduces Vuex as a way to manage shared state between page components. This is needed so that we could pass down the global "loading" state from inertia.js into our page components as needed.

As this store setup is adding more code to our app bootstrapping file: app.ts. This file is restructured for better readability, and to ensure that we are only bootstrapping the inertia app if setup succeeds.

* Use correct button variant and style
* Install latest WiKit pre-release
* Restructure app.ts
* Install Vuex
* Implement store for global loading state
* Add loading to page state

Bug: T290162