Handle server errors when retrieving mismatches
This change redirects any error response by the server back to the home
route and adds an 'errors' flash message to the Inertia response. This
way the Home page can pick up the flash message and indicate a server
error with a Message component, triggered by a computed prop.

Testers will be able to fake/force a server error by sending a
'X-Mismatch-Results-Error' header field with their request to the
results route.

Bug: T289344