Add labels and links (#114)
This change adds an entity ID to Wikidata label map to the results page. This map is then used to display labels for various entity IDs displayed on the page.

Also, as it turns out, `wbformatentitites` only returns a result array with uppercase IDs as keys, however some statement GUIDs are formatted with a lowercase 'q'. Therefore, in this change, we normalize all item IDs to uppercase in the database.

* Add labels to results page
* Add links to Wikidata
* Extract results route handler into ResultsController
* Update browser tests
* Ensure Item Ids are saved in uppercase

Co-authored-by: Silvan Heintze <59574251+Silvan-WMDE@users.noreply.github.com>
Bug: T291569