Handle server errors when storing mismatch reviews
This change redirects any error response when saving mismatch reviews back
to the results page and adds an 'errors' flash message to the Inertia response.
This way the Results page can pick up the flash message and indicate a server
error with a Message component, triggered by a computed prop. Testers will be
able to fake/force a server error by sending a 'X-Mismatch-Finder-Error' header
with the 'mismatch-review' path as value.

The error message with its styles is moved into a new 'MessageBox' component,
so that it can be re-used in and between pages.

Bug: T292055