Add review status column (#118)
This change adds a column to each item's mismatches table, to indicate the review status of the mismatches. Additionally, since there were no tests enabled for the results page, this change adds some missing tests, in preparation for adding more tests in a followup to this PR.

This change does not yet include data binding between the status of the mismatches in the table, and the mismatches in the results pages' model after a new status is selected. These changes will be included in a followup.

* Ensure review status is included in server response
* Add initial unit tests for Results.vue

    This change required us to update some prop definition in the Results page component,
    to make sure that we can mount it in tests.

* Test component status quo
* Add review status column

Bug: T289557