Show only pending mismatches on results page
This change modifies the database query for the results page such, that
only mismatches with review_status 'pending' are retrieved. This however
would lead to the undesired behaviour, that mismatches disappear from
the results page once their review_status has been modified and
submitted, since the inertia PUT request that has been used so far
triggers a new database query and re-renders the page without the
updated mismatch entry.

Therfore, the PUT request to submit review decisions is now sent to the
endpoint as an asynchronous axios request in the background and returns
status code 200 upon success, instead of a redirect back to the results
page.

Bug: T293730