Add GitHub Action config for staging deployment
This adds a GitHub action to deploy Mismatch Finder to the staging
environment on toolforge. Very similarly to the production config, it
first uploads the entire repo directory to a user's toolforge account,
and then syncs to the tool's account from there. The difference to
production is the tool being used ('mismatch-finder-staging') and the
staging directory name in the user's home directory, which is
'mismatch-finder-repo-next' here, instead of 'mismatch-finder-repo',
which is the one for production.

Staging deployments will be run on every push to main, as well as on
branches prefixed by 'staging' or 'gh-actions'.

Bug: T295585