Add website footer (#217)
This change adds a footer to the Mismatch Finder tool's website. The footer is created as a base abstract component, which is meant to be copied and reused in additional wikidata tools in the future. However, it is also able to accept arbitrary context related html which are more relevant to the tool that includes it.

Bug: T287359