Prevent submitting reverted decisions
This change modifies the 'decisions' object in the frontend's Results
page in such a way that it can keep track of the previously submitted
review decision. This way the send() method can ignore submission
requests, when a review decision has been reverted or no decision has
been made at all.

Bug: T297015