Track open vs reviewed mismatches via statsv (#283)
This change adds a scheduled task that runs once per day. It sends the
total number of mismatches and the number of reviewed mismatches to
statsv for reporting purposes. It also adds a deployment.yaml file which
is needed to create a kubernetes cronjob on toolforge, to execute artisan's
schedule:run command once every minute.

Bug: T295760