Add method to parse multiple data values at once
This will be used to parse and format “time”-type values in the results
view; by parsing multiple values at once, we can speed up the display.
(It would be nice to batch-parse during import validation as well, but
that will require more changes.)

There are now two methods to “parse values”: parseValuesForProperty()
directly makes the API call (and is tested via methodProvider()), while
parseValues() batches calls and assembles results (and is tested via
dedicated test methods). This resembles the existing split between
formatEntities() and getLabels(), or getEntities() and
getPropertyDatatypes().