Add sensible default port .env.example
Instead of letting many new developers run into errors while following
the "Quickstart" guide, set a default that has a good chance to work out
of the box.

Furthermore, if that APP_PORT env variable is available, it should be
part of the example env file in either way, because that file gives the
impression of being comprehensive.