Add "missing" review_status (#474)
The code part of this is mostly straightforward – I mainly searched for
occurrences of “both” to find all references to mismatch statuses and
added “missing” directly after the “wikidata” option.

The migration is more painful. Laravel doesn’t support altering ENUM
columns at all. In MySQL, we can issue a raw ALTER TABLE command, adding
the new enum value (at the end of the enum, which is efficient since
MySQL 5.1); in SQLite there’s nothing equivalent (it’s apparently just a
string under the hood? but we still need to tell Laravel that the column
now has an extra supported value), but we only support SQLite for
testing, so just drop the entire table and recreate it from scratch.

Bug: T323204
Co-Authored-By: Michael Große <michael.grosse@wikimedia.de>