Update vertical spacing
Remove the breakpoint-dependent spacing of the header, and instead
always make it $dimension-layout-small. I can’t find any reason for the
margin depending on the background (it was added in aee8dc75cfe), nor a
record of Design asking for this, and we don’t want it now.

Add a margin to the top of the “back” button on the results page; this
makes its total spacing (including the header margin) match that of the
first heading on the home page (as headings have a default top margin).

Remove the top margin of the “which items should be checked?” heading on
the home page, so that the total space between it and the preceding
section is just the section’s own margin, matching the space between
sections on the results page.

Bug: T323678