ImportReport: Optimize database queries
Previously, we were loading the full data of every mismatch in the
mismatches table, and counting the mismatches per review status in
memory. Also, we were querying the mismatches one import at a time,
burdening the databases with a lot of random accesses (based on the
index that exists on the import_id due to the foreign ID constraint)
when really we want the data of all mismatches in the end.

Instead, have the database group and count the mismatches by import and
status, and only send those counts over the wire. This should already be
more efficient (only one full table scan, and less data transferred); it
can optionally be optimized further later, by adding a covering index
for the import_id and review_status, if we need it.