Update sail version in composer.json
Apparently this is the version we already use in composer.lock anyways.
(Anything higher than 1.19 runs into PHP version issues.)

I think this is the last composer change for now – everything else in
`composer outdated` is not a direct dependency and can’t be updated
because of other packages’ requirements.

Bug: T338222