Reassign css to proper div to remove white space below footer (#832)
This wrapper div with the website class was collapsed into the <div id="app">. With the Vue3 rendering this is not happening anymore so the styles need to be reapplied directly to the div app. We need to disable no-multiple-template-root rule in project for this to work, since now there is more than one base element in the template but this error will be gone once the linting rules are set to Vue3 since this rule doesn't exist in Vue3. It will be migrated soon.

Bug: T347190